pyproject.toml
[project]
name = "bedrock-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.11.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "boto3 (>=1.37.29,<2.0.0)",
    "termcolor (>=3.0.1,<4.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "nbstripout (>=0.8.1,<0.9.0)",
]

[tool.poetry]
packages = [{include = "bedrock_agent", from = "src"}]


[tool.poetry.group.dev.dependencies]
notebook = "^7.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
